backend/app/core/auth.py
import os
import bcrypt # Add this
from datetime import datetime, timedelta
from jose import JWTError, jwt
from passlib.context import CryptContext

# 🔧 Fix for passlib + bcrypt version mismatch
if not hasattr(bcrypt, "__about__"):
    class BcryptAbout:
        def __init__(self):
            self.__version__ = bcrypt.__version__
    bcrypt.__about__ = BcryptAbout()

# CONFIG
SECRET_KEY = os.getenv("SECRET_KEY", "ids-super-secret-key-12345")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60 * 24 # 1 Day

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password):
    return pwd_context.hash(password)

def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
