requirements.txt
fastapi
uvicorn
scapy
pandas
numpy
scikit-learn
sqlalchemy
joblib
pydantic
python-dotenv
requests
matplotlib
seaborn
motor
psutil
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
websockets
ipinfo
torch
cryptography
fpdf
